apps/knowledge/serializers.py
from rest_framework import serializers

from .models import Quote, Article


class QuoteSerializer(serializers.ModelSerializer):

    class Meta:
        model = Quote
        fields = ('id', 'author', 'text')


class ArticleSerializer(serializers.ModelSerializer):

    class Meta:
        model = Article
        fields = ('id', 'url', 'title')
