requirements.txt
blessings==1.6
boto==2.34.0
bpython==0.14.2
coverage==3.7.1
curtsies==0.1.19
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.8
django-appconf==0.6
django-bootstrap3==5.0.3
django-braces==1.8.0
django-ckeditor==4.4.6
django-compressor==1.4
django-crispy-forms==1.4.0
django-debug-toolbar==1.3.0
django-enum==0.1
django-enumfield==1.2.1
django-extensions==1.5.5
django-grappelli==2.6.1
django-libsass==0.2
django-model-utils==2.3.1
django-storages==1.1.8
django-toolbelt==0.0.1
django-werkzeug-debugger-runserver==0.2
djangorestframework==3.2.2
drf-nested-routers==0.9.0
enum34==1.0.4
flake8==2.4.0
google-api-python-client==1.4.1
gunicorn==18.0
httplib2==0.9.1
ipaddress==1.0.14
libsass==0.7.0
model-mommy==1.2.4
ndg-httpsclient==0.4.0
oauth2client==1.4.12
oauthlib==0.7.2
pep8==1.6.2
Pillow==2.4.0
psycopg2==2.5.2
pyflakes==0.8.1
Pygments==2.0.2
PyJWT==1.1.0
pyOpenSSL==0.15.1
pystache==0.5.4
python-openid==2.2.5
python-social-auth==0.2.7
requests==2.2.1
requests-oauthlib==0.4.2
rsa==3.2
selenium==2.45.0
simplejson==3.8.0
six==1.9.0
sqlparse==0.1.15
static==1.0.2
static3==0.5.1
uritemplate==0.6
Werkzeug==0.10.4
